Added test.

+++ include/vnix/rat/normalized-pair.hpp
@@ -51,14 +51,14 @@ public:
     if (d == 0) {
       throw "null denominator"; // Do not allow division by zero.
     }
-    enum { NMAX = 1 << (NMR_BITS - 1) };
+    enum { NMAX = uint64_t(1) << (NMR_BITS - 1) };
     if (n_ >= NMAX) {
       throw "numerator too large and positive";
     }
     if (n_ < -NMAX) {
       throw "numerator too large and negative";
     }
-    if (d_ > (1 << DNM_BITS)) {
+    if (d_ > (uint64_t(1) << DNM_BITS)) {
       throw "denominator too large";
     }
   }
